+++ Controls/x264Control.vb
@@ -107,15 +107,15 @@ Public Class x264Control
         components = New System.ComponentModel.Container()
 
         QualityDefinitions = New List(Of QualityItem) From {
-            New QualityItem(16, "Super High", "Super high quality and file size (-crf 16)"),
-            New QualityItem(17, "Very High", "Very high quality and file size (-crf 17)"),
-            New QualityItem(18, "Higher", "Higher quality and file size (-crf 18)"),
-            New QualityItem(19, "High", "High quality and file size (-crf 19)"),
+            New QualityItem(12, "Super High", "Super high quality and file size (-crf 12)"),
+            New QualityItem(14, "Very High", "Very high quality and file size (-crf 14)"),
+            New QualityItem(16, "Higher", "Higher quality and file size (-crf 16)"),
+            New QualityItem(18, "High", "High quality and file size (-crf 18)"),
             New QualityItem(20, "Medium", "Medium quality and file size (-crf 20)"),
-            New QualityItem(21, "Low", "Low quality and file size (-crf 21)"),
-            New QualityItem(22, "Lower", "Lower quality and file size (-crf 22)"),
-            New QualityItem(23, "Very Low", "Very low quality and file size (-crf 23)"),
-            New QualityItem(24, "Super Low", "Super low quality and file size (-rf 24)")}
+            New QualityItem(22, "Low", "Low quality and file size (-crf 22)"),
+            New QualityItem(24, "Lower", "Lower quality and file size (-crf 24)"),
+            New QualityItem(26, "Very Low", "Very low quality and file size (-crf 26)"),
+            New QualityItem(28, "Super Low", "Super low quality and file size (-crf 28)")}
 
         Encoder = enc
         Params = Encoder.Params
@@ -176,6 +176,12 @@ Public Class x264Control
                         cms.Items.Add(New ActionMenuItem(
                                       Params.Tune.Options(x), Sub() SetTune(temp)) With {.Font = If(Params.Tune.Value = x, New Font(Font.FontFamily, 9 * s.UIScaleFactor, FontStyle.Bold), New Font(Font.FontFamily, 9 * s.UIScaleFactor))})
                     Next
+                Case 3 - offset
+                    For x = 0 To Params.Depth.Options.Length - 1
+                        Dim temp = x
+                        cms.Items.Add(New ActionMenuItem(
+                                      Params.Depth.Options(x), Sub() SetDepth(temp)) With {.Font = If(Params.Depth.Value = x, New Font(Font.FontFamily, 9 * s.UIScaleFactor, FontStyle.Bold), New Font(Font.FontFamily, 9 * s.UIScaleFactor))})
+                    Next
             End Select
         End If
     End Sub
@@ -210,6 +216,16 @@ Public Class x264Control
         UpdateControls()
     End Sub
 
+    Sub SetDepth(value As Integer)
+        Dim offset = If(Params.Mode.Value = x264RateMode.Quality, 0, 1)
+        Params.Depth.Value = value
+        Params.ApplyValues(True)
+        Params.ApplyValues(False)
+        lv.Items(3 - offset).SubItems(1).Text = value.ToString
+        lv.Items(3 - offset).Selected = False
+        UpdateControls()
+    End Sub
+
     Function GetQualityCaption(value As Double) As String
         For Each i In QualityDefinitions
             If i.Value = value Then Return value & " - " + i.Text
@@ -219,15 +235,17 @@ Public Class x264Control
     End Function
 
     Sub UpdateControls()
-        If Params.Mode.Value = x264RateMode.Quality AndAlso lv.Items.Count < 4 Then
+        If Params.Mode.Value = x264RateMode.Quality AndAlso lv.Items.Count < 5 Then
             lv.Items.Clear()
             lv.Items.Add(New ListViewItem({"Quality", GetQualityCaption(Params.Quant.Value)}))
             lv.Items.Add(New ListViewItem({"Preset", Params.Preset.OptionText}))
             lv.Items.Add(New ListViewItem({"Tune", Params.Tune.OptionText}))
-        ElseIf Params.Mode.Value <> 2 AndAlso lv.Items.Count <> 3 Then
+            lv.Items.Add(New ListViewItem({"Depth", Params.Depth.OptionText}))
+        ElseIf Params.Mode.Value <> 2 AndAlso lv.Items.Count <> 4 Then
             lv.Items.Clear()
             lv.Items.Add(New ListViewItem({"Preset", Params.Preset.OptionText}))
             lv.Items.Add(New ListViewItem({"Tune", Params.Tune.OptionText}))
+            lv.Items.Add(New ListViewItem({"Depth", Params.Depth.OptionText}))
         End If
 
         Dim offset = If(Params.Mode.Value = x264RateMode.Quality, 0, 1)